pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "argentina-inflation-analysis"
version = "0.1.0"
description = "Pipeline automático para datos de inflación en Argentina"
readme = "README.md"
requires-python = ">=3.9"
authors = [
    {name = "Tu Nombre", email = "tuemail@ejemplo.com"},
]
license = {text = "MIT"}
dependencies = [
    "pandas>=1.5.0",
    "requests>=2.28.0",
    "python-dotenv>=1.0.0",
    "matplotlib>=3.6.0",
    "seaborn>=0.12.0",
    "pytest>=7.0.0",
    "tabulate>=0.9.0",
    "openpyxl>=3.0.0",
    
]

[project.optional-dependencies]
dev = ["black", "flake8", "pre-commit"]
dashboard = ["streamlit>=1.25.0", "plotly>=5.0.0"]